dev-requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-stubs==1.16.0
django-stubs-ext==0.8.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-stubs==1.10.0
drf-yasg==1.21.5
flake8==6.0.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy==1.1.1
mypy-extensions==1.0.0
packaging==23.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
python-decouple==3.8
pytz==2023.2
redis==4.5.3
requests==2.28.2
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
types-pytz==2023.2.0.1
types-PyYAML==6.0.12.9
types-requests==2.28.11.17
types-urllib3==1.26.25.9
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
